pyproject.toml
[project]
name = "excel-model"
version = "0.1.0"
description = "YAML-driven Excel financial model generator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Matthias Christenson" },
]
keywords = ["excel", "financial-model", "openpyxl", "yaml", "p&l", "dcf", "scenario"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "openpyxl>=3.1,<4",
    "polars>=1.0,<2",
    "pyyaml>=6.0,<7",
    "strictyaml>=1.3,<2",
    "dacite>=1.8,<2",
    "click>=8.0,<9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9", "hypothesis>=6.0,<7", "pytest-cov>=5.0,<6"]

[project.urls]
Documentation = "https://neuralsignal.github.io/excel-model/"
Repository = "https://github.com/neuralsignal/excel-model"

[project.scripts]
excel-model = "excel_model.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["excel_model"]

[tool.hatch.build.targets.sdist]
include = ["excel_model/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
